% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMSdetection.R
\name{annotateMSdetection}
\alias{annotateMSdetection}
\title{annotateMSdetection}
\usage{
annotateMSdetection(
  IGSeq_resultset = seq_r06_fasta,
  DIA_resultset = ms_r06,
  write_table = TRUE,
  n_prot_uniqueness_threshold = 1
)
}
\arguments{
\item{IGSeq_resultset}{IgSeq resultset to be annotated with MS detection}

\item{DIA_resultset}{DIA MS resultset based on which to annotate MS detection of the clones in IGSeq_resultset.}

\item{write_table}{Whether to write .csv table of MS detection. Default: TRUE}

\item{n_prot_uniqueness_threshold}{Uniqueness threshold to define clones detected with sufficient specificity.
This may be useful to quantify clonal families with high homology with a certain degree of sharedness of the detectable peptide set.
Defaults to 1, i.e. only peptides mapping to only one of the clones in the context of the current .fasta will be maintained for
 unique analysis}
}
\value{
IGSeq_resultset with added element step7_clonesMSdetection with tables $clones_detected
and quantitative matrices across the MS runs in the DIA_resultset calculated from all mapped or only unique peptides
according to the n_prot_uniqueness threshold set as parameter:
$MS_abundance_allprecursors_mean, $MS_abundance_uniqueprecursors_mean.
In addition, the study design of the DIA_resultset is used to aggregate and report quantities and frequency of observation
(mean and sd of precursor intensities and n_rep) across conditions. There are calculated
including only *unique* precursors (note n_prot_uniqueness_threshold parameter to modify this inclusion criterion).
The result matrices are stored as list elements in the result object with names:
 MS_abundance_uniqueprecursors_mean_cond_rep, MS_abundance_uniqueprecursors_mean_cond,
 MS_abundance_uniqueprecursors_sd_cond, MS_detection_uniqueprecursors_nrep_cond,
 MS_abundance_uniqueprecursors_mean_long. Further, the study_design from the DIA_resultset is propagated to the resultset in slot
 $study_design
}
\description{
annotate MS detection of IGSeq clones based on DIA_resultset
}
