% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleFastaDb.R
\name{assembleFastaDb}
\alias{assembleFastaDb}
\title{assembleFastaDb}
\usage{
assembleFastaDb(
  IGSeq_resultset,
  dataset_tag = NULL,
  write_fasta = TRUE,
  write_fasta_per_sample = FALSE,
  write_clone_tables = TRUE,
  write_fasta_subset_cloneIdGlobal = NULL
)
}
\arguments{
\item{IGSeq_resultset}{IgSeq resultset imported by importIgSeq function.}

\item{dataset_tag}{Dataset tag introduced to cloneIdGlobal and output files. Default: Input IGSeq R object name}

\item{write_fasta}{Whether to write .fasta files to working directory. Default: TRUE}

\item{write_fasta_per_sample}{Whether to write separate .fasta files per sample. Default: FALSE}

\item{write_clone_tables}{Whether to write clone .tsv tables next to .fasta files. Default: TRUE}

\item{subset_cloneIdGlobal}{Whether to subset clones to be included in fasta file. Character vector of global clone IDs.
Default: Null, All clones included.}
}
\value{
IGSeq_resultset with added element step6_clonestofasta
}
\description{
assemble fasta database of variable domains from MiXCR IgSeq results
}
