% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAbundanceTestMS.R
\name{differentialAbundanceTestMS}
\alias{differentialAbundanceTestMS}
\title{differentialAbundanceTestMS}
\usage{
differentialAbundanceTestMS(
  DIA_resultset,
  study_design_external = NULL,
  comparison_matrix = matrix(c("AP_m11_tF", "AP_m11_t0", "AP_m12_tF", "AP_m12_t0",
    "AP_m1_tF", "AP_m1_t0", "AP_m2_tF", "AP_m2_t0"), ncol = 2, byrow = TRUE),
  write_preprocessing_results = TRUE,
  write_diffTable_csv_protein = TRUE,
  write_diffTable_csv_precursor = TRUE,
  overview_label_prefix = "IGSeq_M1imm_APMS_batch1_t3t0",
  plot_pdf = TRUE,
  plot_html = TRUE,
  prot_highlight_tag = "IGSeq"
)
}
\arguments{
\item{DIA_resultset}{DIA-MS resultset imported by importDIAresults function.}

\item{study_design_external}{Alternative study design overwriting the study design in DIA_resultset containing columns filename}

\item{comparison_matrix}{A matrix defining the comparisons to be run.
Column one contains the condition in the counter of each comparison, column two condition in the denominator of the comparison.
Positive log2FCs will mean higher signal in counter condition listed in column 1.
For possible values, check unique(DIA_resultset$study_design$condition).
The data will be subset to the conditions specified in the comparison matrix.}

\item{write_preprocessing_results}{Whether the intermediate results of each pairwise comparison shall be written. Separate subfolders will be generated. Default: FALSE}

\item{overview_label_prefix}{Prefix label for the differential abundance overview .pdf and interactive .html plots}

\item{plot_pdf}{Whether to plot summary of the differential tests.}

\item{plot_html}{If plotting summary, whether to also generate interactive summary of the differential tests.}

\item{prot_highlight_tag}{tag that determines which protein.groups get highlighted}

\item{write_csv_protein}{Whether to write differential abundance testing result table summarized to protein level (csv). Default: TRUE}

\item{write_csv_precursor}{Whether to write differential abundance testing result table containing precursor level information (csv). Default: FALSE}
}
\value{
DIA_resultset with appended element $differentialAbundanceTestResults  with tables
$diffTestRes_prec
$diffTestRes_prot
$comparison_matrix
$data_source
$study_design
$data_long
$data_matrix_log2
$data_matrix_log2_imp
}
\description{
test differential abundance of proteins, clones and clone groups in DIA-MS data on precursor level
}
