% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDIAresults.R
\name{importDIAresults}
\alias{importDIAresults}
\title{importDIAresults}
\usage{
importDIAresults(report_location, study_design = NULL)
}
\arguments{
\item{report_location}{path to the Spectronaut report (Factory default)}
}
\value{
DIA_resultset object (list) with
data_long - data.table of the report
data_wide - data.table of Protein.Group+Precursor.Id over the Ms runs/R.FileName
overview - analysis overview from SN
settings - settings file
}
\description{
import MS results from Spectronaut
}
