% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyCloneIDs.R
\name{modifyCloneIDs}
\alias{modifyCloneIDs}
\title{modifyCloneIDs}
\usage{
modifyCloneIDs(cloneIDs_to_modify, modification_table)
}
\arguments{
\item{cloneIDs_to_modify}{Character vector of protein groups including cloneIDs (with internal | char.)}

\item{modification_table}{modification mapping data table with columns cloneID_old & cloneID_new}
}
\value{
mapping table with input protein groups in column cloneIDs_to_simplify and modified
}
\description{
modify Clone IDs (or protein identifiers) within vector of protein groups.
Useful when non-trivial modifications are required, such as replacement with different information
 from cloneID table.
}
