% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferentialAbundanceOverview.R
\name{plotDifferentialAbundanceOverview}
\alias{plotDifferentialAbundanceOverview}
\title{plotDifferentialAbundanceOverview}
\usage{
plotDifferentialAbundanceOverview(
  diffExpr_result_dt,
  significance_threshold_p = 0.05,
  significance_threshold_fc = 2,
  label_prefix = NULL,
  remove_from_protein_names = "\\\\(Bos;|_MOUSE|_HUMAN",
  browsable_html = TRUE,
  protein_highlight_tag = "IGSeq",
  heatmap = TRUE
)
}
\arguments{
\item{diffExpr_result_dt}{Input data table, typically appended result tables with differential expression testing results as produced by testDifferentialAbundance}

\item{significance_threshold_p}{Significance threshold on p-value to be indicated in plots (p_value_BHadj_protein <= x)}

\item{significance_threshold_fc}{Significance threshold on fold-change to be indicated in plots (|log2FC| >= log2(x)). A two-fold change is 2. The log2FC cutoff is then 1.}

\item{label_prefix}{Prefix label for the output plots. Defaults to "(input R object name)_".}

\item{remove_from_protein_names}{Regular expression of string components to remove from Protein.Names.}

\item{browsable_html}{Whether brosable volcanoe plot .htmls are to be generated. Default: TRUE.}

\item{protein_highlight_tag}{Substring of Protein names that get highlighted (color) in plots.}

\item{heatmap}{Whether an overview heatmap of FCs is to be drawn for the proteins passing the p-value cutoff at least once.}
}
\value{
A (list) containing (access by x$ or by "x[[name]]")
\itemize{
\item diffExpr_result_dt: Reduced diffExpr_result_dt processed for plots
\item volcanos: Volcano plots as ggplot object
\item heatmap: heatmap obj
\item FC_matrix: matrix of Log2FCs of proteins passing the p-value
 significance threshold in any one of the comparisons
}
}
\description{
Produce overview plots of multiple differential test results
}
\author{
Moritz Heusel
}
