% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeIgSeq.R
\name{summarizeIgSeq}
\alias{summarizeIgSeq}
\title{summarizeIgSeqResults}
\usage{
summarizeIgSeq(IGSeq_resultset)
}
\arguments{
\item{IGSeq_resultset}{IgSeq resultset imported by importIgSeq function.}
}
\value{
creates ggplot graphs in out_folder
}
\description{
summarize MiXCR IgSeq results as structured by https://github.com/heuselm/igseq_workflow.git
}
