% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeIgSeqResults.R
\name{summarizeIgSeqResults}
\alias{summarizeIgSeqResults}
\title{summarizeIgSeqResults}
\usage{
summarizeIgSeqResults(result_folder = ".", write_tables = TRUE)
}
\arguments{
\item{result_folder}{Location of the IgSeq-workflow result folder, not yet implemented (needs to be run IN that folder, for now)}

\item{write_tables}{Switch whether the collected summary information should be output as .csv tables}
}
\description{
summarize MiXCR IgSeq results as structured by https://github.com/heuselm/igseq_workflow.git
}
