% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIgSeq.R
\name{importIgSeq}
\alias{importIgSeq}
\title{summarizeIgSeqResults}
\usage{
importIgSeq(
  result_folder = ".",
  write_tables = TRUE,
  report_folder = "importIgSeq_log",
  summarize_in_pdfs = TRUE,
  chain_type_index_end = 4
)
}
\arguments{
\item{result_folder}{Location of the IgSeq-workflow result folder.}

\item{write_tables}{Switch whether the collected summary information should be output as .csv tables}

\item{summarize_in_pdfs}{Whether the results are directly summarized via summarizeIgSeq function. Default: TRUE}

\item{chain_type_index_end}{End character position in AllCHitsWithScore column to assign chain type.
Default: 4 (differentiates isotypes, e.g. IGHG from IGHM). If set to 3, all isotypes will be considered IGH chains, w/o differentiation.}
}
\value{
IGSeq result object (list) with tables
* step1_checkout_log
* step2_histogram_log
* step2_histogram_osq
* step3_assemble_log
* step4_align_log
* step5_cloneAssembly_log"
* step5_finalclones
}
\description{
summarize MiXCR IgSeq results as structured by https://github.com/heuselm/igseq_workflow.git
}
